# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 37AC 识别服务 API
  version: 0.0.1
  description: |
    37AC 角色识别服务的 HTTP 接口定义（Flask 应用 AC_web，源码在 src/server）。
    由源码 + 实测抓包整理，可直接导入 Apifox（导入数据 -> OpenAPI/Swagger）。

    约定与注意事项：
    - 所有 JSON 响应都带 success 与 code（机器可读错误码，2026-09-13 起）。历史结构一律保留：
      多数接口是 {success,message,data}；dashboard/nodes/upload/tasks 是 {type,timestamp,data...}；
      dashboard 异常分支是 {error}（另外补一份 message）。
    - code 取值 = 状态码默认码（OK/CREATED/ACCEPTED/INVALID_PARAMS/UNAUTHORIZED/FORBIDDEN/NOT_FOUND/
      METHOD_NOT_ALLOWED/PAYLOAD_TOO_LARGE/RATE_LIMITED/INTERNAL_ERROR/NOT_IMPLEMENTED/SERVICE_UNAVAILABLE）
      或业务码（AUTH_TOKEN_MISSING/AUTH_TOKEN_INVALID/AUTH_TOKEN_TYPE_INVALID/AUTH_INVALID_CREDENTIALS/
      AUTH_REFRESH_TOKEN_INVALID/PERMISSION_DENIED/API_KEY_MISSING/API_KEY_INVALID/DB_UNAVAILABLE/MAINTENANCE_MODE）。
      完整码表见 docs/api-inventory.md 第 1.1 节。
    - 鉴权分两类：Bearer JWT（用户态接口）与 X-API-Key（识别任务接口）。
    - POST /upload 受滑动窗口限流（默认 5 次/秒/IP），超限 429。
    - 维护模式开启时（settings.maintenance_mode=1），/auth、/admin、/static 之外一律 503。
    - 已知问题详见 docs/api-inventory.md 第 3 节（例如 /tasks/<id> 未校验任务归属）。
servers:
  - url: http://127.0.0.1:13138
    description: 本地/默认（Flask 直连；nginx 8000 只服务 PHP 站点，未反代 API）
...
security: []
paths:
...
  /upload:
    get:
      tags: [识别任务]
      summary: 上传接口使用说明
      responses:
        "200": {description: 成功, content: {application/json: {example: {type: info, message: 37AC 上传 API, usage: {method: POST, url: /upload, headers: {X-API-Key: your_api_key}}}}}}
    post:
      tags: [识别任务]
      summary: 上传图片并识别（排队受理；返回 200，语义上应为 202）
      description: |
        鉴权：X-API-Key。受滑动窗口限流（默认 5 次/秒/IP）-> 429。
        两种响应模式：
        - 默认 JSON：立即返回排队信息，用 task_id 轮询 GET /tasks/{task_id}；
        - 流式：请求头 Accept: text/event-stream 或 X-Stream-Response: true -> SSE 推送 queued/waiting/completed。
        图片来源：multipart 字段 file 或 image；或 JSON/表单字段 image_base64（可带 data URL 前缀）。
      security:
        - apiKeyAuth: []
      parameters:
        - {name: X-Stream-Response, in: header, schema: {type: string, enum: ["true"]}, description: 置 true 走 SSE 流式响应}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: {type: string, format: binary, description: png/jpg/jpeg/webp}
                image: {type: string, format: binary, description: 等价于 file}
                model: {type: string, enum: [37ac, llm, auto], default: 37ac, description: 非法值会被静默改写为 37ac}
          application/json:
            schema:
              type: object
              properties:
                image_base64: {type: string}
                model: {type: string, enum: [37ac, llm, auto]}
      responses:
        "200":
          description: 已受理（JSON 模式）
          content: {application/json: {example: {type: dispatch_task, timestamp: 1789267716, status: queued, message: 图片已上传，等待推理..., task_id: 52213b03-6a5c-48ef-8413-1aac63bf9be8, model: 37ac}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "429": {description: 触发限流, content: {application/json: {example: {success: false, message: 请求过于频繁，请在 1 秒后重试}}}}
  /tasks/{task_id}:
    get:
      tags: [识别任务]
      summary: 查询任务结果（轮询）
      description: 注意：当前实现只校验 API Key 有效性，未校验任务归属（可读取他人任务结果，见 docs/api-inventory.md 3.1）。
      security:
        - apiKeyAuth: []
      parameters:
        - {name: task_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: 任务完成
          content: {application/json: {example: {type: task_result, timestamp: 1789267716, status: completed, message: 任务完成，结果已返回, task_id: 52213b03-..., result: [{name: "原神/荧", prob: 96.06}]}}}
        "202":
          description: 尚未完成
          content: {application/json: {example: {type: task_result, timestamp: 1789267737, status: pending, message: 结果尚未返回, task_id: 52213b03-..., result: []}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {description: 数据库不可用}
  /tasks/{task_id}/stream:
    get:
      tags: [识别任务]
      summary: 任务结果 SSE 流
      description: 浏览器 EventSource 无法自定义 X-API-Key 头，实际使用者是 CLI/服务端；同样未校验任务归属。等待上限 120 秒。
      security:
        - apiKeyAuth: []
      parameters:
        - {name: task_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: SSE 流
          content:
            text/event-stream:
              schema: {type: string}
              example: |
                data: {"status": "completed", "message": "任务已完成", "task_id": "...", "result": [{"name": "原神/荧", "prob": 96.06}], "error": null}
        "401": {$ref: "#/components/responses/Unauthorized"}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: POST /auth/login 返回的 data.access_token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: POST /api-keys 创建，形如 37ac_xxx（仅创建时返回一次）
  responses:
    BadRequest:
      description: 参数错误
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}, example: {success: false, message: 参数错误}}}
    Unauthorized:
      description: 未认证或令牌无效
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}, example: {success: false, message: 未提供认证令牌}}}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success: {type: boolean, example: false}
        code: {type: string, example: INVALID_PARAMS, description: 机器可读错误码，见 api-inventory.md 1.1}
        message: {type: string, example: 参数错误}
...